Attempt at fixing workflow duplicate check.
This is an attempt at fixing the duplicate releases `check_build_id` by implementing a two side operand which checks for the module basename and build_id. Those are also specified on the module file name.

Also added a way for the PIF to loop true known directory's in an attempt to remove issues such as having to copy the pif.json from adb into the modpath of `playintegrityfix`.