fix build module prop
Fixed an issue where beta images don't use name_for_attestation as codename and so won't say it is Beta, So i reverted so it uses vendor name.

Also uncommented a code in extract_images.sh where it deletes the downloaded and extracted images, So that it doesn't clog up your system with images.